[OPS-1447] Document workarounds for running impure tests
Problem: Currently, templates run Haskell test-suites as "checks"
which means that tests are executed as a part of the nix derivation
build, and thus they're pure and cannot interact with some external
resources that may be needed for tests, e.g. DB connection or access
to the Docker socket.

Solution: Document workarounds for running impure tests with haskell.nix.