[OPS-1477] Add cabal check
Problem: In order to publish a package to hackage, it must meet
certain requirements, which can be verified using cabal check. Since
we want to publish our packages to hackage, it makes sense to add a
cabal check step to the CI pipelines of our haskell.nix templates.

Solution: Add cabal check.