[Chore] Document how to set up a haskell project root folder
Problem: The template currently assumes that the haskell project root
is in the same directory as flake.nix, but this is not always the
case. We need to document how to set up a different path to the
haskell project root.

Solution: Document how to set up a haskell project root folder in
flake.nix.