[OPS-1458] Use `makeCI` wrapper
Problem: Our CI template for libraries allows one to specify
ghc-versions so that we can build a library in different
configurations. Currently the only configurable thing is the GHC
version. To do this, we have developed a wrapper over haskell.nix and
want to update our library template to use it.

Solution: Modify our library template to use the `makeCI` wrapper.