[OPS-1531] Update GA pipeline trigger condition to be suitable for forks
Problem: Currently, GA pipeline is only triggered by pushes to branches
within the repo. As a result, CI is never triggered for PRs from forks.

Solution: Use 'on: pull_request:' to allow running CI on PRs from forks.
Use 'on: push: branches: master' to run the pipeline on the default
branch to ensure that checks pass.